+++ frogmouth/screens/main.py
@@ -318,6 +318,18 @@ class Main(Screen[None]):  # pylint:disable=too-many-public-methods
             event.location, remember=event.location != self.query_one(Viewer).location
         )
 
+    def on_history_delete(self, event: History.Delete) -> None:
+        """Handle a request to delete an item from history.
+
+        Args:
+            event: The event to handle.
+        """
+        self.query_one(Viewer).delete_history(event.history_id)
+
+    def on_history_clear(self) -> None:
+        """handle a request to clear down all of history."""
+        self.query_one(Viewer).clear_history()
+
     def on_bookmarks_goto(self, event: Bookmarks.Goto) -> None:
         """Handle a request to go to a bookmark.