+++ frogmouth/widgets/navigation_panes/history.py
@@ -2,28 +2,34 @@
 
 from __future__ import annotations
 
+from functools import partial
 from pathlib import Path
 
 from httpx import URL
 from rich.text import Text
 from textual.app import ComposeResult
+from textual.binding import Binding
 from textual.message import Message
 from textual.widgets import OptionList
 from textual.widgets.option_list import Option
 
+from ...dialogs import YesNoDialog
 from .navigation_pane import NavigationPane
 
 
 class Entry(Option):
     """An entry in the history."""
 
-    def __init__(self, location: Path | URL) -> None:
+    def __init__(self, history_id: int, location: Path | URL) -> None:
         """Initialise the history entry item.
 
         Args:
+            history_id: The ID of the item of history.
             location: The location being added to history.
         """
         super().__init__(self._as_prompt(location))
+        self.history_id = history_id
+        """The ID of the item of history."""
         self.location = location
         """The location for his entry in the history."""
 
@@ -62,6 +68,12 @@ class History(NavigationPane):
     }
     """
 
+    BINDINGS = [
+        Binding("delete", "delete", "Delete the history item"),
+        Binding("backspace", "clear", "Clean the history"),
+    ]
+    """The bindings for the history navigation pane."""
+
     def __init__(self) -> None:
         """Initialise the history navigation pane."""
         super().__init__("History")
@@ -84,8 +96,8 @@ class History(NavigationPane):
         value.
         """
         option_list = self.query_one(OptionList).clear_options()
-        for location in reversed(locations):
-            option_list.add_option(Entry(location))
+        for history_id, location in reversed(list(enumerate(locations))):
+            option_list.add_option(Entry(history_id, location))
 
     class Goto(Message):
         """Message that requests the viewer goes to a given location."""
@@ -109,3 +121,61 @@ class History(NavigationPane):
         event.stop()
         assert isinstance(event.option, Entry)
         self.post_message(self.Goto(event.option.location))
+
+    class Delete(Message):
+        """Message that requests the viewer to delete an item of history."""
+
+        def __init__(self, history_id: int) -> None:
+            """initialise the history delete message.
+
+            args:
+                history_id: The ID of the item of history to delete.
+            """
+            super().__init__()
+            self.history_id = history_id
+            """The ID of the item of history to delete."""
+
+    def delete_history(self, history_id: int, delete_it: bool) -> None:
+        """Delete a given history entry.
+
+        Args:
+            history_id: The ID of the item of history to delete.
+            delete_it: Should it be deleted?
+        """
+        if delete_it:
+            self.post_message(self.Delete(history_id))
+
+    def action_delete(self) -> None:
+        """Delete the highlighted item from history."""
+        history = self.query_one(OptionList)
+        if (item := history.highlighted) is not None:
+            assert isinstance(entry := history.get_option_at_index(item), Entry)
+            self.app.push_screen(
+                YesNoDialog(
+                    "Delete history entry?",
+                    "Are you sure you want to delete the history entry?",
+                ),
+                partial(self.delete_history, entry.history_id),
+            )
+
+    class Clear(Message):
+        """Message that requests that the history be cleared."""
+
+    def clear_history(self, clear_it: bool) -> None:
+        """Perform a history clear.
+
+        Args:
+            clear_it: Should it be cleared?
+        """
+        if clear_it:
+            self.post_message(self.Clear())
+
+    def action_clear(self) -> None:
+        """Clear out the whole history."""
+        self.app.push_screen(
+            YesNoDialog(
+                "Clear history?",
+                "Are you sure you want to clear everything out of history?",
+            ),
+            self.clear_history,
+        )